refactor some layout def.
And starting to put labels in the correct hierarchy. Lots to be done.
It's also now possible to put a box around an entrance label for instance (using -attr style x)

And also it should be possible to unlock and define new label styles in the main 'overview.thl' definition.


Former-commit-id: 7396fa57646cfceb17686e702b445fcc65ee41f2